Remove Rust code
After trying out Ratatui and Textual, it became clear that Textual is a
more high level TUI framework with a lot of cool features build in. This
will make development of a TUI much more efficient compared to Ratatui.

With `uv` there is also a good solution to install a Python tool.

Although building an agent is not a difficult thing to do as it is just
a loop, but Pydantic AI will just take some work off my shoulders. So i
can be more productive in building it.